+++ .pre-commit-config.yaml
@@ -3,6 +3,7 @@ repos:
     rev: 1.8.0
     hooks:
       - id: poetry-check
+        args: [ --lock ]
       - id: poetry-lock
       - id: poetry-export
       - id: poetry-install
