+++ docker/Dockerfile
@@ -1,4 +1,4 @@
-ARG APP_NAME=radio_browser
+ARG APP_NAME=hermes
 ARG APP_PATH=/opt/$APP_NAME
 ARG PYTHON_VERSION=3.13.0
 ARG POETRY_VERSION=1.8.4
