+++ hermes/config.py
@@ -9,14 +9,14 @@ class Config:
 
     # General Config
     ENVIRONMENT = environ.get("ENVIRONMENT", "development")
-    FLASK_APP = environ.get("FLASK_APP", "radio_browser")
+    FLASK_APP = environ.get("FLASK_APP", "hermes")
     FLASK_DEBUG = environ.get("FLASK_DEBUG", False)
     SECRET_KEY = environ.get("SECRET_KEY", "dev")
     STATIC_FOLDER = "static"
     TEMPLATES_FOLDER = "templates"
 
     # Database
-    SQLALCHEMY_DATABASE_URI = environ.get("SQLALCHEMY_DATABASE_URI", "sqlite:///radio_browser.sqlite")
+    SQLALCHEMY_DATABASE_URI = environ.get("SQLALCHEMY_DATABASE_URI", "sqlite:///hermes.sqlite")
     SQLALCHEMY_TRACK_MODIFICATIONS = False
 
     # AWS Secrets
