+++ hermes/enum.py
